PLAT-620 | Extract flashcardItem
