+++ resources/assets/js/components/course/Screen.vue
@@ -23,7 +23,7 @@
 	import End from 'js/components/course/screens/End'
 	import Html from 'js/components/course/screens/Html'
 	import MockExam from 'js/components/course/screens/MockExam'
-	import Flashcards from 'js/components/course/screens/Flashcards'
+	import Flashcards from 'js/components/course/screens/flashcards/Flashcards'
 	import Qna from 'js/components/qna/Qna'
 	import Quiz from 'js/components/quiz/Quiz'
 	import Slideshow from 'js/components/course/screens/slideshow/Slideshow'
