+++ resources/assets/js/components/course/screens/flashcards/Flashcards.vue
@@ -39,64 +39,13 @@
 				ponów cały zestaw
 			</div>
 			<ol class="flashcards-set__list">
-				<li
+				<wnl-flashcard-item
 						v-for="(flashcard, index) in set.flashcards"
 						:key="flashcard.id"
-						:class="['flashcards-list__item', flashcard.answer !== 'unsolved' && 'flashcards-list__item--solved']"
-				>
-					<span class="flashcards-list__item__index">{{index + 1}}</span>
-					<div class="flashcards-list__item__container">
-						<p class="flashcards-list__item__text">{{flashcard.content}}</p>
-						<div class="flashcards-list__item__buttons" v-if="flashcard.answer === 'unsolved'">
-							<a class="flashcards-list__item__buttons__button text--easy"
-							   @click="submitAnswer(flashcard, 'easy')">
-								<span class="icon"><i :class="['fa', ANSWERS_MAP.easy.iconClass]"></i></span>
-								<span class="flashcards-list__item__buttons__button__text">Łatwe</span>
-							</a>
-							<a class="flashcards-list__item__buttons__button text--hard"
-							   @click="submitAnswer(flashcard, 'hard')">
-								<span class="icon"><i :class="['fa', ANSWERS_MAP.hard.iconClass]"></i></span>
-								<span class="flashcards-list__item__buttons__button__text">Trudne</span>
-							</a>
-							<a
-									class="flashcards-list__item__buttons__button text--do-not-know"
-									@click="submitAnswer(flashcard, 'do_not_know')"
-							>
-								<span class="icon"><i :class="['fa', ANSWERS_MAP.do_not_know.iconClass]"></i></span>
-								<span class="flashcards-list__item__buttons__button__text">Nie Wiem</span>
-							</a>
-						</div>
-						<div
-								class="flashcards-list__item__buttons flashcards-list__item__buttons--retake"
-								@click="onRetakeFlashcard(flashcard)"
-								v-else
-						>
-							<span class="flashcards-list__item__buttons__button">
-								<span class="icon"><i class="fa fa-undo"></i></span>
-							</span>
-							<span :class="['flashcards-list__item__buttons__button', ANSWERS_MAP[flashcard.answer].buttonClass]">
-								<span class="icon"><i
-										:class="['fa', ANSWERS_MAP[flashcard.answer].iconClass]"></i></span>
-								<span class="flashcards-list__item__buttons__button__text">{{ANSWERS_MAP[flashcard.answer].text}}</span>
-							</span>
-						</div>
-					</div>
-					<wnl-form
-							method="post"
-							suppressEnter="true"
-							resetAfterSubmit="true"
-							resourceRoute="test"
-							:name="`flashcardNote-${flashcard.id}`"
-							@submitSuccess="onSubmitSuccess">
-						<label class="label">Notatka {{flashcard.id}}</label>
-						<wnl-quill
-								name="note"
-								class="margin bottom"
-								:options="{ theme: 'snow', placeholder: 'Wpisz swoją notatkę...' }"
-								v-model="notes[flashcard.id]"
-						/>
-					</wnl-form>
-				</li>
+						:flashcard="flashcard"
+						:index="index"
+						:context="{type: context, id: screenData.id}"
+				/>
 			</ol>
 		</div>
 		<div class="flashcards-scroll" @click="scrollTop">
@@ -194,81 +143,6 @@
 			&__list
 				margin: $margin-small
 
-		.flashcards-list
-			padding-left: 15px
-
-			&__item
-				display: flex
-				align-items: center
-
-				&--solved
-					color: $color-gray-dimmed
-
-					.flashcards-list__item__container
-						background: $color-background-lightest-gray
-
-				&__index
-					font-weight: $font-weight-bold
-					text-align: right
-
-				&__container
-					display: flex
-					border: $border-light-gray
-					padding: $margin-base
-					flex-grow: 1
-					margin: $margin-small 0 $margin-small $margin-small
-					min-height: 54px
-					flex-direction: column
-
-					@media #{$media-query-tablet}
-						flex-direction: row
-						align-items: center
-
-				&__text
-					flex-grow: 1
-					color: inherit
-
-				&__buttons
-					text-align: right
-					display: flex
-					align-items: center
-
-					@media #{$media-query-tablet}
-						flex: 0 0 $buttonWidth * 3
-
-					&--retake
-						flex: 0 0 $buttonWidth * 2
-
-						.flashcards-list__item__buttons__button .icon .fa-undo
-							font-size: 16px
-
-					&__button
-						opacity: 1
-						display: flex
-						flex-direction: column
-						align-items: center
-						margin: 0 $margin-small
-						cursor: pointer
-
-						@media #{$media-query-tablet}
-							flex-basis: 78px
-							width: 64px
-
-						&:hover
-							opacity: 1
-
-						.icon
-							width: 24px
-							height: 24px
-
-							.fa
-								font-size: 24px
-
-						&__text
-							font-weight: $font-weight-bold
-							text-transform: uppercase
-							font-size: 12px
-
 		.flashcards-scroll
 			width: 32px
 			height: 32px
@@ -293,25 +167,8 @@
 	import {get} from 'lodash';
 	import {scrollToElement} from 'js/utils/animations'
 	import * as mutationsTypes from "js/store/mutations-types";
-	import { Quill as WnlQuill, Form as WnlForm } from 'js/components/global/form';
-
-	const ANSWERS_MAP = {
-		easy: {
-			text: 'Łatwe',
-			iconClass: 'fa-smile-o',
-			buttonClass: 'text--easy'
-		},
-		hard: {
-			text: 'Trudne',
-			iconClass: 'fa-meh-o',
-			buttonClass: 'text--hard'
-		},
-		do_not_know: {
-			text: 'Nie wiem',
-			iconClass: 'fa-frown-o',
-			buttonClass: 'text--do-not-know'
-		}
-	}
+	import WnlFlashcardItem from 'js/components/course/screens/flashcards/FlashcardItem';
+	import {ANSWERS_MAP} from 'js/consts/flashcard';
 
 	export default {
 		props: {
@@ -325,15 +182,12 @@
 			}
 		},
 		components: {
-			WnlQuill,
-			WnlForm,
+			WnlFlashcardItem,
 		},
 		data() {
 			return {
 				ANSWERS_MAP,
 				applicableSetsIds: [],
-				attachedData: {},
-				notes: {},
 			}
 		},
 		computed: {
@@ -355,7 +209,7 @@
 			}
 		},
 		methods: {
-			...mapActions('flashcards', ['setFlashcardsSet', 'postAnswer']),
+			...mapActions('flashcards', ['setFlashcardsSet']),
 			...mapMutations('flashcards', {
 				'updateFlashcard': mutationsTypes.FLASHCARDS_UPDATE_FLASHCARD
 			}),
@@ -365,26 +219,9 @@
 			scrollTop() {
 				scrollToElement(document.getElementById('flashacardsSetHeader'));
 			},
-			onRetakeFlashcard(flashcard) {
-				this.updateFlashcard({
-					...flashcard,
-					answer: 'unsolved'
-				})
-			},
 			onRetakeSet(set) {
 				set.flashcards.forEach(this.onRetakeFlashcard)
 			},
-			async submitAnswer(flashcard, answer) {
-				await this.postAnswer({
-					flashcard,
-					answer,
-					context_type: this.context,
-					context_id: this.screenData.id
-				})
-			},
-			onSubmitSuccess() {
-
-			}
 		},
 		async mounted() {
 			const resources = get(this.screenData, 'meta.resources', []);
