+++ resources/assets/js/consts/flashcard.js
@@ -1,0 +1,17 @@
+export const ANSWERS_MAP = {
+	easy: {
+		text: 'Łatwe',
+		iconClass: 'fa-smile-o',
+		buttonClass: 'text--easy'
+	},
+	hard: {
+		text: 'Trudne',
+		iconClass: 'fa-meh-o',
+		buttonClass: 'text--hard'
+	},
+	do_not_know: {
+		text: 'Nie wiem',
+		iconClass: 'fa-frown-o',
+		buttonClass: 'text--do-not-know'
+	}
+};
